% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_school_enrollment.R
\name{by_school_enrollment}
\alias{by_school_enrollment}
\title{Print School Enrollment By School}
\usage{
by_school_enrollment(x)
}
\arguments{
\item{x}{SCHOOL NAME}
}
\value{
Numeric Vector
}
\description{
Print the total number of school children in each school by name for the 2013-2014 school year separated by gender.
}
\examples{
by_school_enrollment("KS")

by_school_enrollment("NY")

by_school_enrollment("CA")

}
