% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRDCdata.R
\docType{data}
\name{CRDCdata}
\alias{CRDCdata}
\title{Total Enrollment data for 2013-2014 school year}
\format{A data frame with 95507 rows and 66 variables}
\usage{
data(CRDCdata)
}
\description{
Total Enrollment data for 2013-2014 school year
}
\references{
\url{https://usedgov.github.io/api/crdc.html}
}
\author{
Kendall Loh \email{kendall.loh@columbia.edu}
}
\keyword{datasets}
