% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_Function.R
\name{GET_CRDCdata}
\alias{GET_CRDCdata}
\alias{CRDC_api}
\title{GET request to API for CRDCdata
Imports the most up-to-date dataset from the CRDC website and creates a useable dataframe.
A static version of the dataset is also available as part of this pacakge.}
\usage{
CRDC_api(api_key)
}
\arguments{
\item{api_key}{The unique API Key assigned to each user. Please sign up for one here: "https://usedgov.github.io/key/"}
}
\value{
A useable dataframe containing Public School Enrollment data for the 2013-2014 school year. .
}
\description{
GET request to API for CRDCdata
Imports the most up-to-date dataset from the CRDC website and creates a useable dataframe.
A static version of the dataset is also available as part of this pacakge.
}
