% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_table.R
\docType{data}
\name{graph_table-data}
\alias{graph_table-data}
\alias{graph_table}
\title{Enrollment and Census Information}
\format{An object of class \code{data.frame} with 51 rows and 5 columns.}
\usage{
graph_table
}
\description{
Enrollment and Census Information
}
\references{
\url{https://usedgov.github.io/api/crdc.html}
}
\author{
Kendall Loh \email{kendall.loh@columbia.edu}
}
\keyword{datasets}
