% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphit.R
\name{graphit}
\alias{graphit}
\title{Print Graph of X, Y data}
\arguments{
\item{data}{Appropriate dataframe, in this case "graph_table"}

\item{column1}{Selected column within the dataframe, in this case "Percentage"}
}
\value{
graphit
}
\description{
Uses a provides geom_point() graph of a dataframe a \code{y} columns TBD.
}
\examples{
graphit(graph_table, "Percentage")

}
