% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_f_count.R
\name{m_f_count}
\alias{m_f_count}
\title{Print Sum M/F Children Enrolled}
\arguments{
\item{x}{State Abbreviation}
}
\value{
Numeric Vector
}
\description{
Print the total number of school children in a state for the 2013-2014 school year separated by gender.
}
\examples{
m_f_count("KS")

m_f_count("NY")

m_f_count("CA")

}
