% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numschools_state.R
\name{numschools_state}
\alias{numschools_state}
\title{Print Sum Total Schools in State}
\arguments{
\item{x}{State Abbreviation}
}
\value{
Numeric Vector
}
\description{
Print the total number of schools in a state for the 2013-2014 school year.
}
\examples{
numschools_state("KS")

numschools_state("NY")

numschools_state("CA")

}
