% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_data.R
\docType{data}
\name{population_data-data}
\alias{population_data-data}
\alias{population_data}
\title{Census Data}
\format{An object of class \code{data.frame} with 53 rows and 696 columns.}
\usage{
population_data
}
\description{
Census Data
}
\references{
\url{https://usedgov.github.io/api/crdc.html}
}
\author{
Kendall Loh \email{kendall.loh@columbia.edu}
}
\keyword{datasets}
