% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tot_pop.R
\name{tot_pop}
\alias{tot_pop}
\title{Print Sum Total Children Enrolled}
\arguments{
\item{x}{State Abbreviation}
}
\value{
Numeric Vector
}
\description{
Print the total number of school children in a state for the 2013-2014 school year.
}
\examples{
tot_pop("KS")

tot_pop("NY")

tot_pop("CA")

}
